requirements.txt
ccxt
pandas
python-dotenv
numpy
streamlit
plotly
duckdb
